+++ core/lib/model/push/push_config.dart
@@ -1,8 +1,7 @@
-///Push Notification Config
+/// Push Notification Config
 class PushConfig {
-  /** If [shouldDeliverCallbackOnForegroundClick] is true, when push notification
-   * is clicked on app foreground, MoEngage SDK will not handle the redirection.
-   * **/
+  /// If [shouldDeliverCallbackOnForegroundClick] is true, when push notification
+  /// is clicked on app foreground, MoEngage SDK will not handle the redirection.
   bool shouldDeliverCallbackOnForegroundClick;
 
   PushConfig({required this.shouldDeliverCallbackOnForegroundClick});