+++ core/CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ## 5.4.0
 - Android
-    - Support for Foreground Push Click Callback
+  - Compile SDK Version Updated to 33
+  - Native SDK updated to support version `12.8.00` and above.
+  - Support for Foreground Push Click Callback
 - iOS
     - MoEngage-iOS-SDK version updated to `~>9.8.0`