+++ core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # MoEngage Flutter Plugin
 
+# 31-05-2023
+
+## 5.4.0
+- Android
+  - Compile SDK Version Updated to 33
+  - Native SDK updated to support version `12.8.00` and above.
+  - Support for Foreground Push Click Callback
+- iOS
+    - MoEngage-iOS-SDK version updated to `~>9.8.0`
+    
 # 16-05-2023
 
 ## 5.3.1